ziwei_report_year_aliyun_11_12/requirements.txt
aiohttp==3.11.18
aiomysql==0.2.0
cython
fastapi==0.116.0
langchain_community==0.3.27
langchain_core==0.3.68
langchain_openai==0.3.27
lunardate==0.2.2
pandas==2.3.1
prometheus_client==0.22.0
pydantic==2.11.7
python-dotenv==1.1.1
redis==5.2.1
setuptools==75.8.0
tiktoken==0.9.0
urllib3==2.4.0
uvicorn==0.35.0


starlette>=0.40,<1
PyMySQL==1.1.0
cryptography>=41.0.0,<42.0.0
requests>=2.31,<3
zhdate==0.1
#typing-extensions==4.8.0
#charset-normalizer==3.3.2
#certifi==2023.11.17
idna==3.6
#httpx==0.25.2
anyio>=4.0,<5
psutil==5.9.6
openpyxl>=3.1,<4
